마이페이지 댓글 및 디폴트 로그인 계정 세션 삽입
1. MyPage에 있는 게시물 상세보기 클릭시, 댓글 기능 구현완료
2. 로그인이 필요한 페이지에 디폴트 계정(test@naver.com)삽입, 즉 이제 로그인을 하지 않아도 접근 가능

##### 발생 이슈 #####
1. 마이페이지 상세보기에서 내용이 계속해서 누적하여 더해지는 현상 발생 --> 내용은 innerHTML로, 댓글은 init함수를 새로 만들어 해결

#####  To Do #####
1. 마이페이지 상세보기 UI 개선
2. 댓글삭제 작성자만 가능하도록
3. 게시물 의류링크 태그
4. 마이페이지 팔로워, 팔로잉 클릭시 친구리스트 뿌려주기
5. 의류 잡지 크롤링
6. 검색(해시태그 검색 및 친구 검색등)
7. 해시태그 클릭 이벤트
8. 글쓰기 UI 개선
9. 댓글 작성 경과 시간(데이터는 넘겨줌)
10. 댓글 친구 태그
11. 친구 클릭시 친구 피드로 이동(댓글은 해놓음)
12. 의류관련 데이터 구축
13. 친구 언팔로우 양방향 관계 해제

+++ src/main/java/com/team2/webservice/sprint1/service/BoardService.java
@@ -8,7 +8,7 @@ public interface BoardService {
     void write(Board board, String writer);
 
     //게시물 상세보기
-    Board read(Long pid);
+    Board read(int pid);
 
     //게시물 수정
     void update(Board board);
